--- Survinya/Assets/Source/Gameplay/Battle/Scripts.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: aa3b06862768333438ec2a2e010c1167
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
