--- Survinya/Assets/Source/Gameplay/Battle/Scripts/Mono.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 0405e859a327a574385d6ef07bfea9b9
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
