+++ Survinya/Assets/Source/GameplayTest.meta
@@ -1,6 +1,7 @@
 fileFormatVersion: 2
-guid: 78f1ac74011d89643b72abf65ab2dbad
-AssemblyDefinitionImporter:
+guid: e41852f5183a07b4a99d69831cad3137
+folderAsset: yes
+DefaultImporter:
   externalObjects: {}
   userData: 
   assetBundleName:
